+++ internal/config/config.go
@@ -47,6 +47,7 @@ type Config struct {
 	}
 	General struct {
 		AdditionalLogs []LogConfig `yaml:"additional_logs"`
+		DropOldLogs    *bool       `yaml:"drop_old_logs"`
 	}
 }
 
@@ -210,6 +211,13 @@ func validateConfig(config *Config) bool {
 	}
 
 	config.General.AdditionalLogs = validLogs
+
+	// If the cleanup flag is not set, default to true
+	if config.General.DropOldLogs == nil {
+		log.Println("drop_old_logs is not set, defaulting to true")
+		defaultCleanup := true
+		config.General.DropOldLogs = &defaultCleanup
+	}
 
 	return true
 }